+++ public/worker/world.mjs
@@ -97,7 +97,6 @@ const world_belief = {
 const world_mind = new Cosmos.Mind('world');
 const state = world_mind.create_state(1);
 state.add_beliefs(world_belief);
-state.lock();
 
 const player = DB.get_belief_by_label('player');
 if (!player) throw new Error('Player belief not found');
@@ -105,8 +104,8 @@ const player_mind = player.get_trait(state, 'mind');
 let player_state = [...player_mind.state][0];
 player_state = player_state.branch_state(state);
 player_state.learn_about(DB.get_belief_by_label('hammer'), ['location']);
-player_state.lock();
 
+state.lock();
 log(player_state);
 
 //const ball = state.add_belief({